phase2_indexing/requirements.txt
faiss-cpu==1.9.0
numpy==2.2.3
sentence-transformers==3.4.1
